c++: Fix bug in algorithm that caused fewer explorations
When a read had more than one write it could read from, it only
read from one of them, while pushing the rest of them to the Stack.

However, the new form of the algorithm where the recursive explorations
of stores happen during the visit to a StoreInst, may cause
the revisit set to be shrinked during a recursive call, thus changing
the expected behaviour of the algorithm (since the revisit set will
not be the same when popping a <w, r> pair as the time this pair
was pushed.

This commis fixed this problem by pushing the current revisit set
to the stack when pushing a <w, r> pair.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>