c++: Code refactor - part 2
A big disadvantage of the C++ implementation so far is that it
depends (e.g., by storing information about the current execution,
or invoking it to check whether and RMW succeeded)
on the LLVM-IR Interpreter more than it should. This commit is a
step forward on decoupling the classes relevant to the execution
graph with the class of the interpreter.

Problem that will hopefully be solved:
    - Circular dependencies between the Interpreter and the
      ExecutionGraph class.

Hoped-for benefits:
    - Code will become clearer.
    - The execution context of each thread will not be stored in
      the Interpreter; this well relieve us from the need to store
      the execution context before recursive calls. If the execution
      context of each thread is stored in the graph, we will not
      have to copy the stack frames of all threads before performing
      a recursive call.
    - Better code performance.
    - Better memory handling and no memory leaks.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>