c++: Fix loop unrolling
Until now, when a loop needed to be unrolled, the body of the loop
was copied N times (where N is the unroll value), and if the loop
had not terminated after N iterations, some IR transformations
forced the code to jump to a special block containing a
__VERIFIER_assume(false) statement.

However, that caused programs with infinite loops
(e.g., of the form: for (;;) do_something();) not to be handled
properly, since the condition of the assume statement was never
met, and all the executions were dropped.

This commit fixes this issue by replacing the call to
__VERIFIER_assume(false) with a call to the newly-created
and internally-declared function __end_loop().