c++: Fix bug occurring in buggy SC executions
When an assertion is violated in an execution, we have to make sure that
psc is acyclic before reporting the violation. Just returning if psc
returns a cycle is not enough though (duh), as one has to also clear
the execution stacks of all threads and before continuing. This commit
fixes this issue.