c++: Implement Thread-Local Storage (TLS)
This commit adds support for variables declared as thread-local.
Until now, these were treated as plain global variables which
was not correct.

This commit adds an unordered_map for each thread where the values
of all the thread-local variables are stored. At each new execution,
these maps are re-initialized with the initial values for the
thread-local variables.

Of course, this comes at a price (adds runtime overhead) but is
nevertheless a useful feature.