c++: Implement Thread-Local-Storage for main()
It is useful to be able to manipulate thread-local variables
from the main() function, mostly for initialization purposes.
This was not possible until now, since main() was treated differently
from the other threads. The only caveat is that all thread-local
variables must be zero-initialized (bad things will happen if they
are not).