c++: Add default cases for Event's overloaded <<
There are constructors (e.g., the EventLabel constructor for
thread-starting events) that do not initialize all the fields of
the EventLabel. Obviously, if we try to fetch the value for such
a field, we will get garbage.

This is the case -getting garbage values- when printing an EventLabel
w/o optimizations enabled: the program tried to print, e.g.,
an ordering with a garbage value, and since none of the matching cases
was triggered, control was returned to the caller without returning
a value (undefined behavior). To avoid this undefined behavior, we can
add a default guard in ovefrloaded operators that returns the stream,
instead of not handling such cases at all.