tests: Add a buggy RMW testcase
It seems the way RMWs are currently handled is buggy. More
specifically, one cannot postpone the addition of the write event (W1)
of a read (R1) that is now exclusive until R1 is encountered again,
since we may encounter a new RMW (R2, W2) that needs to read the
value written by (R1, W1), but that value is not present!

This means that once a read is revisited and is now exclusive,
its corresponding write needs to be added immediately. This
testcase illustrates this very fact.