c++: RMW bug fixes + removal of -wb option
This commit introduces changes which can be divided into
three major categories:

a) Fixes the bug described in 4076887eb5
b) Fixes a bug related to the ordering of RMWs
c) Removes -wb config option

---
Commit 4076887eb5 describes a -henceforth fixed- bug with the
handling of RMWs. More specifically, the previous approach
involved revisiting reads, but not immediately adding any missing
store events that are part of a read that, when revisited, becomes
exclusive. As a shortcoming, testcases like RMWFix (named rmw-bug
in commit 4076887eb5) resulted in segmentation faults.

This commit fixes this issue by immediately adding the missing
store when a read that is revisited becomes exclusive,
revisiting all possible reads due to the addition of the store,
and also continuing with the exploration of the resulting graph,
if the graph is consistent. This approach led to significant
performance gains in testcases involving RMWs.

---
In addition, this commit fixes an unrelated bug regarding the
ordering of RMW events. Previously, both ACQUIRE and RELEASE
orderings for FAIs and CASs were treated as ACQ_REL, which is
obviously wrong. Now, the model checker does distinguish between
these different access modes.

This had some interesting consequences for some testcases, e.g.,
MP+incMP. In this testcase, the ordering for the FAIs does matter
as it affects the hb-view of the events. This testcase was
-by mistake- written with the FAIs having ACQUIRE ordering, which
now yielded different results from the expected ones (as the
expected corresponded to ACQ_REL ordering). Thus, this testcase
is fixed, in order to match the expected (ACQ_REL) results. A
relevant litmus test is added to demonstrate the abovementioned.

---
Lastly, the currently unsupported -wb option was removed.