tests: Move szymanski benchmark from cooking/ to correct/
There was not any particular reason this testcase was in cooking,
so move it.

This testcase can be run under Nidhugg-{SC,TSO,PSO} and Nidhugg-POWER
if -DNIGHUGG or -DNIDHUGG_POWER are enabled, respectively. And
since Nidhugg does count the executions where an assume() statement
is not true (whereas RCMC ignores them), if
-DNIDHUGG_PRINT_COMPLETE_EXECS is defined(*), a printf statement is
triggered for executions that are explored in full (there are only
6 of those). Ignoring the executions where some assume() is not
satisfied, RCMC and Nidhugg explore the same number of traces (6).

(*) For SC, TSO and PSO only.