c++,ocaml: Fix bug in calculation of hb;po in both implementation
Consider the following program (small0):

		a := x || b := x || x := 2
		       || x := 1 ||

and that, after adding a couple of events, we have the following graph:

			Wna

		Rx1     Rx0
		---     ---

			Wx1

We arrive to this graph by using Ori's version of the algorithm,
but let us suppose that this is a valid state anyway.

After adding the second write to 'x' we will get some graphs, but
with a particular modification order, suppose [Wx1, Wx2]. However,
after these subexplorations are done, we need to return to the
graph above and try the alternative modification order, i.e.,
[Wx2, Wx1].

Then, in the getRevisitLoadsNonMaximal() function, we discard some
reads that should *not* be revisited, according to some consistency
checks. Here we are interested in the reads discarded due to the
coherence condition (Alg.5, L.4). The previous version of the
getHbPoBefore()/calc_hbpo_before() function erroneously marked the
event 'e' in its argument as it were [hb;po]-before itself, which
is not true (consider the read (a) to be the function's argument),
and in the example above this led to not revisiting the read (a),
thus missing the execution where the reads read 2 and 0, respectively,
and the modification order is [Wx2, Wx1].

This commit fixes this issue by amending getHbPoBefore()/calc_hbpo_before()
appropriately.