c++: Fix PSC calculation for new algorithm
The current (new) version of the algorithm implemented does
not rely on MO, but rather on WB. However, MO is necessary
for consistency checking in the presence of SC accesses
(PSC calculation). This commit modifies PSC calculation
accordingly, with all possible extensions of WB calculated
when checking consistency of an execution.

In addition, two more weak versions of PSC checking (and the
respective command-line options) are introduced:

    1) PSC-weak: where only one extension of WB is checked
       per location.

    2) PSC-wb: where wb is used in the calculation of PSC
       in place of MO.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>