c++: First step towards a version that tracks MO
This commit adds a new version of the algorithm that also
tracks the modification order (MO).

To this end, the RCMCdriver class is made abstract and can
be instantiated as a driver that either tracks MO or WB
(a version that tracks WeakRA should also be implemented
at some point).

In addition, test scripts and expected results have been
modified accordingly.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>