c++: Merge visitStore() and visitRMWStore()
Instead of having two different functions for handling
plain and RMW writes and signaling the driver whenever
each of these events is encountered, this commit merges
these two functions into one VisitStore() function
(similar to the one presented in the paper).

In order to do this, we need also to ensure that the
collection of reads to be revisited will behave correctly
irrespective of the type of the store. Of course, this is
done for both the WB and the MO version of the driver.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>