c++: Fix handling of joins
This commit updates the calculation of rf and views for thread joins.
Although no bug has occured because of this (yet), the previous handling
of views was responsible for the erroneous behavior of the
-count-duplicate-execs switch (failing to count duplicates in some cases).

Joins are a special case of event because their rf (a Finish event),
may not be present in the graph when they are added. When the
respective Finish event is added, the rf of a join change, and the
views were updated.

However, this can potentially lead to bugs since restricting a graph
to a certain event may remove a Finish event, but not its parent join!
Thus, it could be the case that a join had an rf that was not in the
graph something that, although did not create directly a problem,
led the views of the join to be outdated; in turn, the outdated
views led to different string representations of the graph, which
caused the -count-duplicate-execs switch to misbehave.

This commit addresses this issue, by reseting join events when a
graph is cut.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>