c++: Fixes for wb calculation (not complete)
This commit addresses some issues the current calculation of wb
had, which led to the exploration of inconsistent graphs in some
cases. One such example is the inc+inc+RR+W+RR testcase, under
some randomized schedulings. Others are the ones added with
this commit.

While this commit fixes a lot of cases where inconsistent graphs
were explored, it does not fix all of them. For example, the test case
wb2 is still a problematic one; special care needs to be taken
when revisiting events as well, as the current checks are not enough.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>