execgraph,mo,drivermo: Simplify and move splitLocMOBefore()
The splitLocMOBefore() function used to return a pair of indices
indicating which stores of a location's coherence order are
concurrent. However, one index within the mo should suffice,
since all the stores that are mo-after the index will also be
concurrent. This commit addresses this issue, and moves the
relevant function to the MO Driver, since it is only relevant
there.

Signed-off-by: Viktor Vafeiadis <viktor@mpi-sws.org>
Signed-off-by: Michalis Kokologiannakis <michalis@mpi-sws.org>