interp: Add support for __VERIFIER_nondet_int()
This commit adds support for limited data non-determinism
by introducing the __VERIFIER_nondet_int(), the specifications
of which can be found at:

    https://sv-comp.sosy-lab.org/2019/rules.php

Intuitively, each call to __VERIFIER_nondet_int() will return
an integer that is "unpredictable" before the analysis. However,
__VERIFIER_nondet_int() will return the same (per thread)
sequence of numbers in all the explored executions.

Note that using the Mersenne Twister engine for random number
generation makes the code extremely slow, as merely seeding the
engine at the beginning of each execution has a linear
complexity on the state_size. On the other hand, using a simple
multiplicative congruential generator has a constant seeding
complexity, and is thus preferred.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>