conf: Fix clang detection for LLVM-7 and above
The m4 script that sets the variables CLANG and CLANGXX assumed
that the binary for clang is always of the form
clang-"$LLVMVERSION". However, while the "$LLVMVERSION" (obtained
by the respective llvm-config) is always of the form
"MAJOR_VERSION.MINOR_VERSION.REVISION", from LLVM-7 and above,
only the major version is kept in the name of clang binary,
in contrast to previous versions that kept the minor version as
well.

The build procedure was working so far because if the binary
clang-"$LLVMVERSION" is not detected, the script falls back to
the default clang, that happened to be clang-7, thus no
discrepancy between the clang and the llvm version was detected.
In other systems though, such a discrepancy can occur and
terminate the build procedure, even though the correct clang
version is installed.

This commit adds a check in the m4 file that enables it to look
for the correct clang binary.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>