execgraph: Improve RMW chain traversal in WB calculation
The calculation of WB required the calculation of the upper
and lower limit of the RMW chain in which a store belongs to.
This was done repeatedly for each pair of stores, increasing
the complexity of the WB calculation.

This commit precalculates such limits for all stores, which
can be then used in the calculation of WB.

Signed-off-by: Viktor Vafeiadis <viktor@mpi-sws.org>
Reviewed-by: Michalis Kokologiannakis <michalis@mpi-sws.org>