conf,llvm: Change loop bounding mechanism
This commit changes the way loops are bounded, in order for the
respective pass to be able to work after mem2reg. It also runs
loop bounding after mem2reg and spin-assume, which alters the
expected results for some test cases.

Until now loops were "unrolled", by copying the loop body N times,
where N is the loop bound. Since LLVM-IR is in SSA form, the
copied bodies had differently named values, as expected. However,
if mem2reg had ran before the unrolling, it could be the case
that values defined in the loop were used after the loop as well.
In that case, if one of the cloned bodies branched into one of these
uses, the original value (which is a register) would be used
instead of the cloned one, leading to different results.

(Note that a bug in the loop unrolling mechanism with nested loops
existed as well, where the last BB of an unrolled inner loop was
not copied when the body of the outer loop was unrolled. To see
this, run the example in issue #1, with loop unrolling running
after mem2reg.)

Thus, instead of repairing the unrolling method described above,
a counter is used for each loop which is set to N when we enter
a loop, and decreased after each iteration until it reaches 0.
When that happens a special function is called to signal to the
interpreter that this loop should stop.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>