REFACTOR: Make EventLabel abstract and create derived classes
This commit decomposes the EventLabel class to different
subclasses, depending on the type of the action recorded in
the ExecutionGraph.

Apart from speed benefits, this commit also makes the different
types of labels clearly distinct, thus minimizing the chances
for errors in the future (e.g., utilizing fields unused in for
a particular label type).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>