driver*,interp*: Fully decouple Interpreter and ExecutionGraph
This commit removes the last few interactions between the
Interpreter and the ExecutionGraph. The API of the Driver is
redesigned so that the interpreter never has to query the graph
for information; all the necessary information is provided by
the driver when the latter is called from the interpreter.

The code of the driver is refactored, and some getter methods
are introduced as a byproduct of this commit.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>