event,interp,driver,parser: Modify printing in DOT format
Instead of only outputting the relevant source-code line in
DOT printing, it also makes sense to show the actual value
read and the actual location accessed.

This commit modifies DOT printing accordingly, and also removes
the printing of reads-from edges that come from the INIT event,
as that induced too much noise in the past.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>