interp,driver: Fix values returned to the interpreter
This commit changes the values returned to the interpreter from
the driver (see troep.c), and fixes potential bugs in test cases
with CAS instructions.

The problem stemmed from a type mismatch when reading from a
store following a CAS operation. CASes are always performed on
atomic variables (i64 in LLVM-IR). Whenever a CAS fails, the
value read is written in the second argument, and the
produced store has also an i64 type -- same as the CAS.

However, when a subsequent read reads from that location, the
type of the read may be different (can be the actual type of that
variable), and this type is what the interpreter expects. Thus,
we need to make sure we return the expected type when there is a
mismatch between what the read expects and what the write produces.

Reported-by: Carlo Wood <carlo@alinoe.com>
Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>