interp,driver: Properly detect double-free errors
This commit introduces proper detection for double-free errors.
It does so, by looking at the execution graph for existing free
labels, and reports an error if that is the case.

The previous way of keeping an array with all the memory that
had been freed could not detect all such errors, since this
array was populated whenever a new event was added, but then
emptied at the end of each execution. Thus, it would miss such
bugs that would occur at the replay of an execution.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>