execgraph,interp,driver*: Fix bugs in complex RMW test cases
This commit fixes the following three bugs:

    1) casdep on local variables: It could be the case that
an CAS read was dependent on some stack variable that takes
its value from a global read. However, this dependency was not
recorded in the graph as the implementation was not detecting
concurrent accesses on stack variables at the time this branch
was created. This commit makes the minimal changes required for
such a functionality. [Faulty test cases exposing the bug:
STC, QU]

    2) dependence on stack allocas: While adding stack accesses
to the graph is necessary (see above), there need to be a
dependency between the allocation event of a variable and all
accesses of that variable. This commit models this behavior by
introducing a control dependency whenever an alloca is performed
(although a data dependency might have been more correct).
[Faulty test cases exposing the bug: STC]

    3) correctly updating AR view of RMW stores: An RMW store
needs to be able to "see" at least as many things as the last
release of the same thread (same loc), but also all the things
that reads in between the last release (any loc) and the RMW
store are seeing. However, this misses the case where the last
release of the same thread (any loc) is an RMW store itself,
in which case we also need to see whatever the read part of
this RMW store is seeing. [Faulty test cases exposing the bug:
mpmc-queue]

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>