driver*,execgraph: Fix PSC calculation for all models
This commit modifies the PSC calculation in the graph so as it
works for all different models. More specifically, for IMM,
execution validity is now ensured iff (PSC_fence U AR) is
acyclic and, to that end, relevant code that enables the
calculation of specific parts of PSC has been added.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>