driver*,execgraph: Remove separate calculation of PSC_fence
This commit removes the special care that had been taken for
calculating PSC_fence only. This is based two observations:

    1) PSC_base does not add more edges between fences compared
    to the ones added by PSC_fence.
    2) Even for IMM, we need to check the whole PSC for
    consistency anyway.

Thus, we can calculate the full PSC anyway, and simple restrict it
to fence accesses if only fences are of interest. If the point (1)
above turns out to be wrong, this commit can be reverted.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>