driver,interp,pass: Detect accesses to unallocated memory
This commits adds error reporting for cases where memory that
has not been allocated (i.e., not stack/heap/static) is accessed.

To that end, the collection of information regarding variable
names has been improved, and now the names of 2D arrays are
correctly printed.

In addition, the interpreter is now made aware of the
"deallocations" that are happening due to the restriction of a
graph, or the "allocations" that are happening due to the
restoration of a prefix.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>