scripts,event*,execgraph,driver*: Enhance error detection
This commit restores the checking of wrong test cases by
including them again in the test suite, as well as adding
more erroneous test cases.

In addition, it enhances error detection in the following ways:

    1) Accesses to memory that's not currently allocated are
    now detected and properly reported as errors. To this end,
    a new "bottom" event has been introduced in order to avoid
    asking for values that do not exist during replaying/printing.

    2) More memory errors are supported, e.g., double-free,
    deallocation of freed memory, racy malloc()s/free()s, access
    of freed memory, etc, and test cases demonstrating these
    capabilities are added.

Finally, test cases that demonstrate the new printing of variable
names, introduced in #647fc66, are also added.

TODO: Detection of uninitialized locals, and detection of accesses
of deallocated locals should be added as well.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>