driver*,tests: More join() issues fixed
This commits deletes the resetJoin() function (which was not
properly used anyway), and fixes the updating of the views of
join()s. More specifically, it resets _and_ updates the view
of a join() statement when the respective thread finishes, in
a similar style to changeRf() so that the more complicated scheme
where the views of a join() are first reset (when the graph is
restricted), and then updated (when the child terminates) is
abandoned.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>