immdriver: Still fixing the views of join() statements
The second point of commit #09375023 was wrong, in the sense that,
while the PPO view will contain the views of the previous join()
statements, if any of the previous join()s is reset, then the
current join will disappear from the graph as well, since it
was definitely added _after_ all the events of all the previous
joins() were added.

Thus, it is safe to add keep the previous acquires in the PPO
view of a join(), and this commit fixes this issue.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>