tests,interp,driver: Add rmem's tests and fix relevant issues
This commit adds the test cases from Pulte et al.'s paper on
promising semantics for ARMv8/RISC-V. In addition, some minor
issues exposed by these test cases are fixed.

First, a problem with the calculation of dependencies for RMWs
is fixed, where the dependencies of an RMW instruction were
updated between the load and the store part, instead of being
update in the end of the instruction. This led to some problems
exposed by the qu-opt test case.

Second, abort() calls after calls to visitError() have
been deleted. This is because, if visitError() deems some
execution to be inconsistent, then we should continue executing
the remaining threads.

Third, visitError() no longer clears the stack of a thread when
an erroneous execution is found to be inconsistent. This is
because the interpreter might try to update the stack frame of
that thread after the driver returns control back to the
interpreter. Since the driver blocks the thread anyway, we do not
have to clear the stack as well anyway.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>