driver: Prioritize a thread that is acquiring a lock
This commit ensures that, whenever a thread acquires a lock, then
this thread will get prioritized over all other threads, and that
critical sections will open one by one.

Until now, this was not the case (and things seemed to work), but
this seems like a safer and more reasonable choice.

Signed-off-by: Michalis Kokologiannakis <michalis@pinky32.mpi-sws.org>