depexecgraph,interp: Bug fix for prefix saving
This commit fixes an bug (demonstrated by the test case included
in this commit) that caused the tool to explore executions that
violate RMW functionality.

The root of the problem lied in the way successful RMWs are
detected. To find such RMWs we check whether the store part of
the RMW exists in the graph. To see why this creates a problem,
consider what happens when we add a read.

Whenever we are adding a read to the graph, to avoid violating RMW
functionality, we are removing from the possible read-from options
the rfs of successful RMWs. But it can be the case that, after the
restoration of some prefix, only the read part of an RMW is added,
since the store of which the prefix was restored only depends on
reads. Then, an RMW that is bound to be successful (when the store
part is added) is considered unsuccessful, and reading from the
place it reads from is allowed.

This problem was previously fixed in a hacky way, namely by making
events depend on the store part of RMWs as well. But the proper
way to fix this is to include the store part of a successful
RMW in the prefix that is saved, whenever the read part is
included as well. This commit does exactly that.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>