driver: Perform correct consistency checks at each point
Thic commit fixes the way appropriate consistency checks for
a particular program point are calculated.

The problem occurred with LAPOR enabled. Consistency checks needed
to be calculated at each point anyway, but then, to determine
whether we should check for full consistency at a point P, we
only checked whether full consistency needs to be calculated (or
whether we are at an error point), and not whether full consistency
needs to be calculated at P.

This commit addresses this issue, and also creates a hierarchy
of program points such that, if a consistency check needs to be
performed at a point with score N, the check will also be
performed at points with score <= N. (Errors have the lowest
score, so that consistency checks are always performed at errors.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>