driver: Fixes for macOS compatibility
A part of the consistency checks (linear extensions calculation)
was broken for macOS (at least for llvm@6 and llvm@7). This was
due to STL implementation details under Linux and macOS.

More specifically, a part of the code relied on the fact that
different (exact) copies of the same unordered_map<> will be
iterated in the same order. However, it turns out that this is
not the case, probably due to some rehashing taking place in
macOS when an unordered_map<> is copied.

This commit provides a simple fix for the above issue.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>