interp,tests: Restore support for printf() in user programs
One of the previous commits that added support for calculating
dependencies for more instructions, broke support for printf()
(probably most external functions).

That was because the interpreter was attempting to get the last
instruction of the previous stack frame (to check whether that was
a ret inst and update dependencies, even if the last frame was
empty (which could happen if we were calling an external function).

This commit fixes this issue, and also adds a printf() to one
of the existing tests, to make sure that something similar will
not go undetected in the future.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>