driver: Gracefully exit if open() has been called too many times
In cases where open is called too many times, print an error
message. This will be printed even if open() is called many times
sequentially, as the tool doesn't free a file descriptor whenever
close is called (but only when the graph is restricted).

The limit can be tuned by the respective CLI option. Maybe in the
future it's better to simply extend the FD table, instead of
forcing a hard limit.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>