eventlabel,pers,interp,driver*: Redesign persistency handling
This commit includes a redesign of the way persistency guarantees
are enforced, so that the tool accounts for some cases that could
not be handled before (e.g., that dir ops are ordered wrt
subsequent metadata-affecting ops).

To that end, new labels are introduced for different types of
disk writes, and writes no longer maintain the range of other
writes the are ordered with. This should give more flexibility
in case we want to include other persistency models in the future,
or modify the way pb is calculated.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>