driver,enumAPIs: Add routines to check for recovery validity on-the-fly
This commit leverages the relevant CLI options for persistence,
and adds methods that check persistence at each step, if such
an option has been specified by the user (currently it is the
default).

This way a number of redundant explorations for recovery can be
avoided (compared to when checking for validity on error only).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>