driver*,tests: Add a driver with basic CO support for LKMM
This commit adds the boilerplate for LKMM's driver. Most
functionality is missing at this point, but very basic things
(e.g., adding reads/writes to the graph, as well as coherence)
should work.

Coherence is enforced in a different way compared to what LKMM
specifies; hb(c11);eco irreflexivity is checked instead of
po;echo irreflexivity. These two ways should be equivalent, but
it remains to prove that.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>