interp,driver,tests: Fix segfault with RMWs on unallocated space
Having a CAS or a FAI on unallocated space would cause the tool
to crash. There were a handful of reasons why something like that
could happen:

    1) When trying to determine whether a load will become an RMW,
we would ask the interpreter for the value it is reading. Reading
from unallocated space created a problem in such cases.
      - Solution: Check for validity of the address before
        creating the label (thus before checking whether the
        load will become an RMW).

    2) Even if validity is properly checked, we need to ensure
that, when replaying an execution, we will not replay the store
part of the RMW (which belongs in the same LLVM-IR instruction).
Otherwise, reading from BOTTOM would create problems when adding
the store part.
      - Solution: Conditionally execute the store part of the
        RMW, and make sure that the thread gets blocked *also
        during replay*.

    3) Special care needs to be taken for lock CASes, as these
are handled directly in the driver and are not treated as plain
CASes are.
      - Solution: Make sure that solution 2) above also applies
        to locks.

This commit address all the issues above, and also adds relevent
test cases.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>