interp,tests: Correctly handle dependencies of function call args
This commit addresses two issues that have to do with the handling
of dependencies of function call arguments:

    1) The interpreter would segfault when trying to calculate
dependencies of function calls that were not declared.
    2) The interpreter would also segfault when trying to calculate
dependencies of indirect function calls.

This commit fixes both issues by always using the actual (dynamic)
value used for the function call (as opposed to the value
determined statically).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>