config,interp: Do not impose a limit on open files
This commit gets rid of the option used to impose a limit
on the maximum number of (concurrently) open files.

The number of concurrently open files (and given fds) is
hard to determine since unique fds are given even to open()
calls that are not concurrent. Thus, it not as simple as setting
an upper bound on the number of the available fds. This commit
bypasses the problem by not limiting the number of available fds.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>