interp,driver: Introduce variables to track interpreter's state
This commit introduces variables and getters/setters that make
tracking the interpreter's state easier. Presumably, they will
also make it easier to incorporate different memory models that
require manipulation of the interpreter's state (e.g., weakestmo).

Specifically, both the program's state (i.e., main vs recovery)
as well as the execution's state (i.e., normal vs replay) are
tracked.

As a byproduct of this change, we also delegate visitError() to
not replay the execution if we are in a replay, instead of
explicitly checking in the interpreter before calling visitError().

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>