driver: Introduce getDskWriteValue() function
For disk writes, getting the value of the initializer is a special
case, since the location's of a file's data are not explicitly
initialized, and hence getting the value of these location would
result to a segfault.

This commit addresses this problem by creating a getDskWriteValue()
function, which treates the case where DskWrites read from INIT
specially.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>