mdatapass: Fix bug exposed in LLVM-3.8
For some weird reason, the metadata collected for a global
variable differ in LLVM-3.8 when they are collected through
memcpy and through global variable iteration. This creates
entries with the same key (but different value) in the
globalInfo metadata array for some global variables, which
are not removed by std::unique(), leading to infinite
recursion later on.

Since we don't need to collect mdata in memcpy() for global
variables for which we already have metadata, this commit only
collects metadata for a global variable the first time it is
encountered.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>