interp,tests: Model RCU lookups in open()
This commit models the fastpaths of open() by not having open()
calls that do not have the O_CREAT flag not acquire the dir inode
lock. Thus, we get an effect similar to the kernel's dcache. This
also implies that concurrent open calls that do not have O_CREAT
are unordered (this is reflected in cons-op+op).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>