lkmmdriver: Fix ppo definition for LKMM (rfi;dep problem)
Until now, rfi edges were unconditionally included in ppo
in the pporf views of loads. However, this created an issue in
LKMM (see MP+RA), due to the extra ppo edges and the interaction
with prop.

This commit addresses this issue by making sure to include only
the ppo views of when an rf when a read is reading internally.

(This fix should also probably be migrated to IMM.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>