tests: Showing why prop should not be transitive (fix pending)
This commit adds a test which show that prop should not be
transitive. More specifically, if prop is transitive, we get
the edge (1,4) -> (1,3) by transitivity [(1,4) -> (3,3) and
(3,3) -> (1,3)]. In turn, is in prop_i and thus in xb. However,
it contradicts ppo, thus making the execution invalid.

A fix for this is pending.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>