lkmm*: Split xb calculation into parts
This commit splits the calculation of XB into its containing
parts, in preparation of the addition of RCU's calculator. This
is done because, in order to calculate rcu-link, we cannot just
use xb; we need ar and pb separately. Instead of trying to extract
these edges from xb, let's just try to calculate ar and pb
separately for now.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>