include,tests: Rewrite lkmm.h
This commit rewrites a large part of lkmm.h and also fixes some
bugs in the same file. The rewrite is done so as to not rely on
stdatomic.h types at all, and to be able to have both atomic
and non-atomic accesses to the same memory location (which is not
possible when using _Atomic).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>