incl: Fix the definition of barrier in lkmm.h
The previous definition (atomic_signal_fence(memory_order_relaxed))
does not seem to prevent the compiler from reordering two writes
when optimizations are enabled. Perhaps it is better to model
using inline assembly although, if it is deemed necessary to
include such an event in the graph, we will need to take special
care anyway.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>