eventlabel: Do not copy a read label's revisit set
The default copy constructor of a read label would copy the
read label's revisit set as well, which is rather undesirable
for our use case, since the copied labels are normally used in
revisit prefixes. This commit introduces an explicit copy
constructor that avoids this issue.

This change results in approximately 3.5% performance gain
on pinky (197 -> ~190 sec), and to much better memory consumption
in some benchmarks. An example of the benefit in performance
is casw(7) where now genmc needs 11'' instead of 16.6'' that it
used to need. An example of the benefit in memory is lamport(4),
where genmc now consumes 2GB of memory after 13', while it used
to consume 10.5GB after only 2.5'.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>