eventlabel,execgraph,driver: Move revisit set infra to driver
This commit moves the revisit set infrastructure to the driver,
where it is more natural for it to be. This way, we get rid of
the explicit copy constructor for ReadLabels, which does seem
a bit out of place. There might be a small performance penalty,
but this is not obvious/significant so far.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>